ipam/multi-pool: Set v4/v6 pod alloc CIDRs into local node
When in kvstore mode, the agent does not start the CilumNode k8s watcher
for the remote nodes, but only for the local one.  Instead, the agent
watches the kvstore to get updates related to the remote nodes.
Therefore, to make multi-pool IPAM compatible with kvstore mode, the
propagation of the remote node pod CIDRs has to pass through the
kvstore.

The commit sets IPv4 and IPv6 primary and secondary pod allocation CIDRs
into the local node store from the IPAM multi-pool manager following
each update to the CiliumNode allocated CIDRs. In turn, the local node
store will trigger the NodeDiscovery that will ultimately  write the
updated pod CIDRs into the kvstore for the remote nodes to consume.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>